feat(sidebars-mutations): moved out SIDEBAR_TYPE from query
