+++ examples/shop-with-blog/client/src/components/Sidebar/SidebarQuery.js
@@ -11,13 +11,6 @@ export const GET_SIDEBAR_STATE = gql`
   }
 `;
 
-export const SIDEBAR_TYPE = {
-  cart: 'cart',
-  account: 'account',
-  signUp: 'signUp',
-  forgotPassword: 'forgotPassword'
-};
-
 export class SidebarQuery extends Query {
   static defaultProps = {
     query: GET_SIDEBAR_STATE
