+++ examples/shop-with-blog/client/src/components/Sidebar/index.js
@@ -2,3 +2,4 @@ export * from './CloseSidebarMutation';
 export * from './OpenSidebarMutation';
 export * from './SidebarQuery';
 export * from './SidebarContainer';
+export * from './sidebarType';
