+++ examples/shop-with-blog/client/src/components/Sidebar/sidebarType.js
@@ -1,0 +1,6 @@
+export const SIDEBAR_TYPE = {
+  cart: 'cart',
+  account: 'account',
+  signUp: 'signUp',
+  forgotPassword: 'forgotPassword'
+};
